im running out of ways to say i worked on the credits

+++ Slider/Assets/Scripts/Credits/CreditsTile.cs
@@ -15,6 +15,6 @@ public class CreditsTile : MonoBehaviour
     public class DelayedEvent
     {
         public UnityEvent Event;
-        public int Delay;
+        public float Delay;
     }
 }
